Tons of work on reflection, properties, etc.

+++ trunk/runtime/lib/Debug.c
@@ -7,14 +7,14 @@
 #include <stdlib.h>
 
 void Debug_write(const String * msg) {
-  fprintf(stderr, "%.*s", msg->length, msg->start);
+  fprintf(stderr, "%.*s", (int32_t) msg->length, msg->start);
 }
 
 void Debug_writeLn(const String * msg) {
-  fprintf(stderr, "%.*s\n", msg->length, msg->start);
+  fprintf(stderr, "%.*s\n", (int32_t) msg->length, msg->start);
 }
 
 void Debug_fail(const String * msg) {
-  fprintf(stderr, "%.*s\n", msg->length, msg->start);
+  fprintf(stderr, "%.*s\n", (int32_t) msg->length, msg->start);
   exit(1);
 }
